limit the semaphore wait
