+++ src/Amplitude.Net.Tests/BasicTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Runtime.CompilerServices;
-using System.Threading;
 using System.Threading.Tasks;
 using Bogus;
 using Microsoft.Extensions.Configuration;
@@ -144,44 +143,5 @@ public class BasicTests
         }
 
         Test();
-    }
-}
-
-public class Counters
-{
-    public int Calls;
-    public int Successes;
-}
-
-public class MockHttpMessageHandler : HttpClientHandler
-{
-    private readonly Counters _counters;
-    private Exception? _exception;
-
-    public MockHttpMessageHandler(Counters counters)
-    {
-        _counters = counters;
-    }
-
-    public Exception? Exception => _exception;
-
-    protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-    {
-        Interlocked.Increment(ref _counters.Calls);
-        try
-        {
-            var response = await base.SendAsync(request, cancellationToken);
-            if (response.IsSuccessStatusCode)
-            {
-                Interlocked.Increment(ref _counters.Successes);
-            }
-
-            return response;
-        }
-        catch (Exception ex)
-        {
-            _exception = ex;
-            throw;
-        }
     }
 }
