+++ src/Amplitude.Net.Tests/Counters.cs
@@ -1,0 +1,7 @@
+namespace Amplitude.Net.Tests;
+
+public class Counters
+{
+    public int Calls;
+    public int Successes;
+}
